Db, mapper: encrypt/decrypt archive password on-the-fly.
Still it does not feel right to store passwords in a data-base, and only
encrypted with a password which is literally available in the cloud's
config.php. However, this is also what the mail app does and probably
other apps. The problem is that the cron jobs (which run in "anonymous"
mode, unauthenticated) otherwise would not have access to the data.