Finish encryption support.
There is one peculiarity with encrypt zip-files: they can be listed.
Unfortunately, ATM the underlying backend does not provide the
information that the archive is encrypted, it just fails when one tries
to access the data.

At the moment this can be cured with user-interaction by adding a
password in the side-tab. It will be propagated to the database of
existing mounts.