ArchiveStorage::hasUpdated($path, $timestamp) -- fix logic.
hasUpdate() now returns true if the $timestamp argument is earlier than
the mtime of the associated archive-file. This should be the correct
implementation as the $timestamp argument is just the storage_mtime from
the filescache table. Hence, if the archive file has been updated, we
cannot know if stuff is already in-place and should trigger a cache
invalidation.

This also should make sure that no cache shoot-downs occur unless the
archive file has been changed.