Mime-type hacks. This commit is evil because (you should read on ...)
The current state of the Nextcloud server distribution is quite
incomplete concerning the supported MIME-types (otherwise not, of
course).

Also, adding support for yet unsupported MIME-types is really overly
difficult.

The hacks of these commits hack around these may-be-short-comings but
exploiting internals of the NC server implementation.

Oh well. But it just works. :)