MountProvider: shoot down the files-cache on unmount.
Actually keeping the cache entries would not be so bad, but for the
frontend. Unfortunately the path "store" (Vue-sprech) does not listen on
"files:node:deleted" events. Thus unmounting and mounting again give
duplicate directory entries in the front-end.

Shooting down the files-cache makes sure that we get a new file-id on
the next mount and thus the stale frontend-cache entry for the old
file-id does not hurt.

Otherwise we would have to submit a pull-request to NC core.