Fix usage of the file-cache.
As the mount is read-only we just always use the file-cache entries
except for accessing the file-data (which of course must be extracted
from the archive file).

If the scanner is invoked on the mounted archive storage, however, then
we re-scan the archive file.

So during the initial mount the cache is filled. Should it ever be
necessary to rescan then archive, then the archive file is scanned
again.