Remove the success notification from a background mount on move and delete.
Idea is that too many notifications can be annoying and keeping a
success notification after unmount does not make too much sense.

For the rename: maybe it would make sense to keep the notification as
the links contained there just use file-ids and will continue to work
even after rename. Also: maybe once upon a time in the future we may
want to add a remove or unmount action to the notification. Then it
makes sense to not remove it on rename.