Add a "repair" step which installs missing mime types and mime-type aliases.
Caveat: do not install mime-type for already registered extensions, as
this leads to inconsistencies with already registered mime-type in the
databse.

Installing mime-type aliases should in principle fix the display of
generic (gear) symbols, but in practice it shows that this may or may
not work.

The rolw of `occ maintenance:mimetype:update-js` and `occ maintenance:mimetype:update-db`
is somewhat unclear: sometimes the symbols are correctly display without
the occ command, sometimes not. Nextcloud chaos.