Workaround NC core's copy from ro-storage permissions bug.
See

- rotdrop/nextcloud-app-files-archive#55
- nextcloud/server#53726
- nextcloud/server#53733

Nextcloud up to and including version 32 copies the
read-only-permissions of a read-only source to the target of a copy
operation, even if the destination storage is writable.

This should be fixed upstream, as a workaround a forced scan of the
destination seems to also fix the problem at the cost of performance, of
course.